moves the mapCluster method to the PeakCluster class to clean up the code a bit
